refactor: pkg dataapps: split code to more files
