+++ internal/pkg/service/appsproxy/dataapps/config.go
@@ -1,0 +1,89 @@
+package dataapps
+
+import (
+	"context"
+	"time"
+
+	"github.com/keboola/go-client/pkg/request"
+	"github.com/pquerna/cachecontrol/cacheobject"
+)
+
+const (
+	OIDCProvider = ProviderType("oidc")
+	PathPrefix   = RuleType("pathPrefix")
+)
+
+type AppProxyConfig struct {
+	ID             string         `json:"-"`
+	Name           string         `json:"name"`
+	UpstreamAppURL string         `json:"upstreamAppUrl"`
+	AuthProviders  []AuthProvider `json:"authProviders"`
+	AuthRules      []AuthRule     `json:"authRules"`
+	eTag           string
+	maxAge         time.Duration
+}
+
+type AuthProvider struct {
+	ID           string       `json:"id"`
+	Name         string       `json:"name"`
+	Type         ProviderType `json:"type"`
+	ClientID     string       `json:"clientId"`
+	ClientSecret string       `json:"clientSecret"`
+	IssuerURL    string       `json:"issuerUrl"`
+	LogoutURL    string       `json:"logoutUrl"`
+	AllowedRoles *[]string    `json:"allowedRoles"`
+}
+
+type AuthRule struct {
+	Type         RuleType `json:"type"`
+	Value        string   `json:"value"`
+	Auth         []string `json:"auth"`
+	AuthRequired *bool    `json:"authRequired"`
+}
+
+type ProviderType string
+
+type RuleType string
+
+func GetAppProxyConfig(sender request.Sender, appID string, eTag string) request.APIRequest[*AppProxyConfig] {
+	result := &AppProxyConfig{}
+	req := request.NewHTTPRequest(sender).
+		WithError(&SandboxesError{}).
+		WithResult(result).
+		WithGet("apps/{appId}/proxy-config").
+		AndPathParam("appId", appID).
+		AndHeader("If-None-Match", eTag).
+		WithOnSuccess(func(ctx context.Context, response request.HTTPResponse) error {
+			// Add app id to the result
+			result.ID = appID
+
+			// Use provider id as fallback until name is added to Sandboxes API
+			for i, provider := range result.AuthProviders {
+				if provider.Name == "" {
+					result.AuthProviders[i].Name = provider.ID
+				}
+			}
+
+			// Add ETag to result
+			result.eTag = response.ResponseHeader().Get("ETag")
+
+			// Process Cache-Control header
+			cacheControl := response.ResponseHeader().Get("Cache-Control")
+			if cacheControl == "" {
+				return nil
+			}
+
+			cacheDirectives, err := cacheobject.ParseResponseCacheControl(cacheControl)
+			if err != nil {
+				return err
+			}
+
+			if cacheDirectives.NoStore || cacheDirectives.NoCache != nil {
+				return nil
+			}
+
+			result.maxAge = time.Second * time.Duration(cacheDirectives.MaxAge)
+			return nil
+		})
+	return request.NewAPIRequest(result, req)
+}
